refactor:  DQN agents to implement Dueling DQN architecture

+++ core/env/observations.py
@@ -7,10 +7,9 @@ if TYPE_CHECKING:
 from core.env.types import DIR_OFFSETS, Direction, GridType
 
 
-def observe_vec11(env: "SnakeEnv") -> npt.NDArray[np.uint8]:
-    # Numpy boolean still uses 1 byte, use uint8 to be unified type with more complex observations
+def observe_vec11(env: "SnakeEnv") -> npt.NDArray[np.float32]:
     if not env.snake:
-        return np.zeros(11, dtype=np.uint8)
+        return np.zeros(11, dtype=np.float32)
     r, c = env.snake.body[0]
     dy, dx = DIR_OFFSETS[env.snake.dir]
     y, x = next(iter(env.apples)) if env.apples else (r, c)
@@ -37,7 +36,7 @@ def observe_vec11(env: "SnakeEnv") -> npt.NDArray[np.uint8]:
             y < r,
             y > r,
         ],
-        dtype=np.uint8,
+        dtype=np.float32,
     )
